requirements.txt
pandas>=2.1.4
prophet
matplotlib
numpy>=1.26.0
scikit-learn
seaborn
datasets>=2.17.1
huggingface_hub
python-dotenv>=1.0.0
statsmodels
scipy>=1.11.3
torch>=2.1,<2.5
transformers
xgboost
chronos-forecasting
gluonts>=0.14.3
git+https://github.com/SalesforceAIResearch/uni2ts.git
einops>=0.7.0
hydra-core>=1.3.0
omegaconf>=2.2
jax[cpu]
jaxtyping>=0.2.24
lightning>=2.0
tensorboard
